pyproject.toml
[project]
name = "datasamorybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "python-telegram-bot>=22.1",
    "anthropic>=0.40.0",
    "aiohttp>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "newspaper3k>=0.2.8",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "lxml[html-clean]>=5.4.0",
]
